Fix: Resolve Vercel build errors and warnings
This commit addresses several issues that were causing the Vercel build to fail or produce warnings:

1.  **`date-fns` Locale Import:** Corrected the Farsi (fa) locale import in `components/forms/TaskFormModal.tsx`. The import path was changed from `date-fns/locale` to `date-fns/locale/fa` as per `date-fns` v2+ standards. This resolves the "Attempted import error: 'fa' is not exported from 'date-fns/locale'".

2.  **`themeColor` Metadata:** Removed `themeColor` from the `metadata` object in `app/layout.tsx`. Next.js recommends using the `viewport` export for `themeColor` configuration, which was already in place. This resolves the "Unsupported metadata themeColor" warning.

3.  **`sharp` Build Scripts:** Added `sharp` to `pnpm.allowedVideoHostProviders` in `package.json`. This is to address the warning about `sharp` build scripts being ignored, which could potentially cause issues if `sharp` is used for image processing.

Addressed issues identified in the Vercel build log, aiming to achieve a successful deployment.
