+++ package.json
@@ -66,7 +66,7 @@
     "next-pwa": "latest",
     "next-themes": "^0.4.4",
     "react": "^19",
-    "react-day-picker": "8.10.1",
+    "react-day-picker": "9.7.0",
     "react-dom": "^19",
     "react-hook-form": "^7.54.1",
     "react-native": "latest",
@@ -78,7 +78,7 @@
     "tailwindcss-animate": "^1.0.7",
     "three": "latest",
     "use-sync-external-store": "latest",
-    "vaul": "^0.9.6",
+    "vaul": "^1.1.2",
     "zod": "^3.24.1",
     "zustand": "latest"
   },
@@ -89,5 +89,10 @@
     "postcss": "^8.5",
     "tailwindcss": "^3.4.17",
     "typescript": "^5"
+  },
+  "pnpm": {
+    "allowedVideoHostProviders": [
+      "sharp"
+    ]
   }
 }
